+++ demisto_sdk/commands/common/tests/script_test.py
@@ -504,3 +504,36 @@ class TestScriptValidator:
         validator = ScriptValidator(structure_validator)
 
         assert not validator.check_separators_in_files()
+
+    DEPRECATED_VALID = {"deprecated": True, "comment": "Deprecated. Use the XXXX script instead."}
+    DEPRECATED_VALID2 = {"deprecated": True, "comment": "Deprecated. Feodo Tracker no longer supports this feed "
+                                                        "No available replacement."}
+    DEPRECATED_VALID3 = {"deprecated": True, "comment": "Deprecated. The script uses an unsupported scraping "
+                                                        "API. Use Proofpoint Protection Server v2 script instead."}
+    DEPRECATED_INVALID_DESC = {"deprecated": True, "comment": "Deprecated."}
+    DEPRECATED_INVALID_DESC2 = {"deprecated": True, "comment": "Use the ServiceNow script to manage..."}
+    DEPRECATED_INVALID_DESC3 = {"deprecated": True, "comment": "Deprecated. The script uses an unsupported scraping"
+                                                               " API."}
+    DEPRECATED_INPUTS = [
+        (DEPRECATED_VALID, True),
+        (DEPRECATED_VALID2, True),
+        (DEPRECATED_VALID3, True),
+        (DEPRECATED_INVALID_DESC, False),
+        (DEPRECATED_INVALID_DESC2, False),
+        (DEPRECATED_INVALID_DESC3, False)
+    ]
+
+    @pytest.mark.parametrize("current, answer", DEPRECATED_INPUTS)
+    def test_is_valid_deprecated_script(self, current, answer):
+        """
+        Given
+            - A deprecated script with a description.
+
+        When
+            - running is_valid_as_deprecated.
+
+        Then
+            - a script with an invalid description will be errored.
+        """
+        validator = get_validator(current_file=current)
+        assert validator.is_valid_as_deprecated() is answer