+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 # 1.3.8
 * Updated the **secrets** command to work on forked branches.
 * Added a validation verifying that the pack's README.md file is not equal to pack description.
+* Improved the deprecated message validation required from playbooks and scripts.
 
 # 1.3.7
 * Added a validation to ensure correct image and description file names.